redefined style for all pages. Added Unspash API to get random images.
Used the bootstrap cards for listing the all the campsites as well as for showing individual sites. Reformatted the new and edit campground pages. Added some fields for camprounds such as price, image and description. Installed the Unsplash package in order to get images from the Unsplash API. Added the node-fetch@v2 package in order to get the Unspash API working.