added a utility folder. Created a error handler class file and an catch async error function wrapper.
Created the utils folder which contains the ExpressError.js and catchAsync.js files. ExpressError.js is a basic error class wich contains a message and a status code. The catchAsync.js file is a wrapper function that is able to catch errors that async functions throw. Modified all the async protocols in app.js so that they use the catchAsync wrapper. This new feature has been tested and seems to work.